+++ src/others/components/Layout.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { ReactNode } from "react";
 import Box from "@mui/material/Box";
 
-import { DashboardContextProvider } from "./dashboard-context";
+import { SidebarContextProvider } from "./sidebar-context";
 
 export interface LayoutProps {
   header?: ReactNode;
@@ -10,11 +10,11 @@ export interface LayoutProps {
 
 export const Layout: React.FunctionComponent<LayoutProps> = ({ header, children }) => {
   return (
-    <DashboardContextProvider>
+    <SidebarContextProvider>
       <Box sx={{ height: "100vh" }}>
         {header}
         <Box sx={{ height: "calc(100vh - 64px)" }}>{children}</Box>
       </Box>
-    </DashboardContextProvider>
+    </SidebarContextProvider>
   );
 };
